bili/bili-aggregator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
PyYAML==6.0.2
feedparser==6.0.11
requests==2.32.3
